+++ DevelopGitHub/GitHubExtension/GitHubExtension.Statistics.WebApi/Services/Implementations/StatisticsService.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using GitHubExtension.Statistics.WebApi.CommunicationModels;
 using GitHubExtension.Statistics.WebApi.Models;
 
-namespace GitHubExtension.Statistics.WebApi.Services
+namespace GitHubExtension.Statistics.WebApi.Services.Implementations
 {
     public class StatisticsService : IStatisticsService
     {
